aws-greengrass/services/detector/requirements.txt
flask
pillow
awsiotsdk
